django_authentication/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-cors-headers
django-environ
gunicorn
